requirements-dev.txt
attrs >=17.4
black ==22.12.0
bump2version >=1.0.1
check-manifest >=0.36
compas_invocations
doc8
flake8
invoke >=0.14
isort
sphinx_compas2_theme
twine
wheel
jupyter
-e .
